+++ src/diagram/ProbabilisticOperator.cpp
@@ -21,7 +21,7 @@ ProbabilisticOperator::ProbabilisticOperator(const std::string &name, const std:
 
 DeltaQ ProbabilisticOperator::calculateDeltaQ(const System &system, const DeltaQ &deltaQ)
 {
-    std::vector<double> resultingCdf;
+    std::vector<long double> resultingCdf;
     std::vector<DeltaQ> scaledDeltaQs;
     /*
     for (const auto &entry : followingComponentAndProbabilities) {
